Design
I have looked at many robotic hand designs and decided that this one would be a good design. I think this because it can easly fit 4 sevos in the arm for the fingers and one in the hand for the tumb. It uses flexible paracord to keep the finges and thumb up right