Design Change
I have decided to change the design of my robotic arm. I changed it from having 5 servos, one for each finger. To having only two servos, one for four fingers and one for the thumb. I decided to do this because of the cos of my project. The original design would have cost up to 150 euros and that was a lot more than I was willing to spend on it. The new design costs only 50 euros which is around what I was wanting to spend in the first place. The problem with only using two servos instead of four is that the hand will have less movement, this is because all fingers except for the thumb are moved by one servo so they will all be moved at the same time.